♻️(tooling) move agent_factory to identite
