+++ src/web/tests/infrastructure/identite/test_create_agent.py
@@ -6,8 +6,8 @@ from config.app_config import AppConfig
 from domain.identite.errors.agent_errors import ProfilAgentAlreadyExists
 from infrastructure.di.identite.identite_container import IdentiteContainer
 from infrastructure.gateways.shared.logger import LoggerService
+from tests.factories.identite.agent_factory import AgentFactory
 from tests.factories.identite.utilisateur_factory import UtilisateurFactory
-from tests.factories.recruteur.agent_factory import AgentFactory
 
 fake = Faker()
 
@@ -27,7 +27,7 @@ def test_create_agent(identite_integration_container):
         email=fake.email(),
         prenom=fake.first_name(),
         nom=fake.last_name(),
-        matricule=fake.bothify("MAT-####"),
+        intitule_poste=fake.bothify("MAT-####"),
     )
 
     result = identite_integration_container.create_agent_usecase().execute(input_data)
@@ -35,7 +35,7 @@ def test_create_agent(identite_integration_container):
     assert result.email == input_data.email
     assert result.prenom == input_data.prenom
     assert result.nom == input_data.nom
-    assert result.matricule == input_data.matricule
+    assert result.intitule_poste == input_data.intitule_poste
 
 
 def test_create_agent_with_existing_user(identite_integration_container):
@@ -44,7 +44,7 @@ def test_create_agent_with_existing_user(identite_integration_container):
         email=existing_user.email,
         prenom=fake.first_name(),
         nom=fake.last_name(),
-        matricule=fake.bothify("MAT-####"),
+        intitule_poste=fake.bothify("MAT-####"),
     )
 
     result = identite_integration_container.create_agent_usecase().execute(input_data)
@@ -58,7 +58,7 @@ def test_cannot_create_agent_twice(identite_integration_container):
         email=existing_agent.utilisateur.email,
         prenom=fake.first_name(),
         nom=fake.last_name(),
-        matricule=fake.bothify("MAT-####"),
+        intitule_poste=fake.bothify("MAT-####"),
     )
 
     with pytest.raises(ProfilAgentAlreadyExists):
